Added Semantic-UI CSS framework and created basic page styling with state updates onChange
Semantic-UI styling documentation can be found here: https://react.semantic-ui.com/

The static page now has inputs to allow the user to input the following fields:

-Starting Value
-Starting Unit
-Converted Value
-Converted Unit

After a change on any of the form fields local state will be updated and the checkSuccess function will
run.

Next steps are to have the checkSuccess function run another pure function which will return a response
that will update state with isSuccess = true and then will render in the GUI to tell the user the calc-
ulation is correct.